Remove unneessary casts
Following https://github.com/knative/serving/pull/7689 remove the casts of the consts.
Consts are 'number' type and are autocased by compiler.

+++ pkg/reconciler/autoscaling/resources/sks_test.go
@@ -56,7 +56,7 @@ func TestMakeSKS(t *testing.T) {
 	}
 
 	const mode = nv1a1.SKSOperationModeServe
-	const na = int32(42)
+	const na = 42
 
 	want := &nv1a1.ServerlessService{
 		ObjectMeta: metav1.ObjectMeta{
